Complete algorithm
The algorithm is complete, maybe some function can be written in a more readable way + it can definitely be written in a more performing way.
Nonetheless, it has all the pieces and works. Unfortunately, it does not consistently beat the heuristic. Maybe a deeper NN would do the trick.
After training, the NN evaluates the agent on famous instances.